requirements.txt
# azure-mgmt-resource>=18.0.0
azure-identity>=1.5.0
azure-mgmt-containerregistry
azure-mgmt-containerinstance
azure-containerregistry
python-dotenv
aiohttp==3.7.4
azure-core==1.24.0
azure-mgmt-core
azure-mgmt-resource==20.1.0
docker